In UserProfileResponse.cs replaced the int fields with long, as the COUNT(*) result in mariadb returns that data type and there was a failure when using int.

+++ Core/DTOs/UserProfileResponse.cs
@@ -1,4 +1,4 @@
 namespace Core.DTOs;
 
-public record UserProfileResponse(string Username, string Email, DateTime CreatedAt, int CreatedPlaylistsCount,
-    int PendingTracksCount, int ApprovedTracksCount);
+public record UserProfileResponse(string Username, string Email, DateTime CreatedAt, long CreatedPlaylistsCount,
+    long PendingTracksCount, long ApprovedTracksCount);
